Remove stack feature
On second thought, implementing something like the stack requires implementing many other aspects of the game engine, namely triggered and activated abilities. Players can potentially be given a way to manually put such effects on the stack, but the user experience will be quite clunky for something that's not super valuable. All of this can be accomplished via the chat.